fix: SdkKeyInfo 로 컴포넌트 변경
