+++ src/features/merchant/api-keys/components/SdkKeyInfo.tsx
@@ -1,0 +1,84 @@
+import { sdkApi } from '@/api/sdk/api';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Switch } from '@/components/ui/switch';
+import * as styles from '@/features/merchant/api-keys/styles/sdk-key.css';
+import { useCopy } from '@/hooks/useCopy';
+import { useMutation } from '@tanstack/react-query';
+import { Copy } from 'lucide-react';
+import { useState } from 'react';
+import { toast } from 'sonner';
+
+interface SdkKeyInfoProps {
+  sdkKey: string;
+}
+
+const SdkKeyInfo = ({ sdkKey }: SdkKeyInfoProps) => {
+  const handleCopy = useCopy();
+
+  const [isActive, setIsActive] = useState(true);
+
+  const { mutate: activateSdkKey } = useMutation({
+    mutationFn: () => sdkApi.activate(),
+    onSuccess: () => {
+      setIsActive(true);
+      toast.success('SDK 키가 활성화되었습니다.');
+    },
+  });
+
+  const { mutate: deactivateSdkKey } = useMutation({
+    mutationFn: () => sdkApi.deactivate(),
+    onSuccess: () => {
+      setIsActive(false);
+      toast.success('SDK 키가 비활성화되었습니다.');
+    },
+  });
+
+  const handleToggleActive = (checked: boolean) => {
+    if (checked) {
+      activateSdkKey();
+    } else {
+      deactivateSdkKey();
+    }
+  };
+
+  return (
+    <div className={styles.content}>
+      <div className={styles.keySection}>
+        <Label>SDK 키</Label>
+        <div className={styles.sdkKeyContent}>
+          <Input
+            value={sdkKey}
+            readOnly
+            suffix={
+              <Button
+                variant="ghost"
+                size="sm"
+                onClick={() => handleCopy({ key: sdkKey, message: 'SDK 키가 복사되었습니다.' })}
+                disabled={!isActive}
+              >
+                <Copy size={16} />
+              </Button>
+            }
+            disabled={!isActive}
+            className={styles.input}
+          />
+          <div className={styles.statusWrapper}>
+            <div className={`${styles.statusBadge} ${isActive ? styles.active : styles.inactive}`}>
+              {isActive ? '활성' : '비활성'}
+            </div>
+            <Switch
+              size={'lg'}
+              checked={isActive}
+              onCheckedChange={handleToggleActive}
+              aria-label="SDK 키 상태 토글"
+            />
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default SdkKeyInfo;
